LSTM
The dataset consists of 10929 instances and 561 features, which is captured using an accelerometer and a gyroscope. The data is taken from 3-axial linear acceleration and 3-axial angular velocity data for 30 people performing static, dynamic and postural transitions . A Neural Network is built with 2 LSTM cells to classify the kind of activity a person is performing.